// src/app/_enums/worplan-category.enum.ts
export enum WorkplanCategory {
  RH_DIRECT = 'RH_DIRECT',
  RH_INDIRECT = 'RH_INDIRECT',
  TRIP = 'TRIP',
  TRAINING = 'TRAINING',
  SERVICE_TECHNOLOGY = 'SERVICE_TECHNOLOGY',
  // SERVICE_OTHER = 'SERVICE_OTHER',
  EQUIPMENT_TECHNOLOGY = 'EQUIPMENT_TECHNOLOGY',
  // EQUIPMENT_OTHER = 'EQUIPMENT_OTHER',
  SOFTWARE_LICENSES = 'SOFTWARE_LICENSES',
  EQUIPMENT_AND_SOFTWARE = 'EQUIPMENT_AND_SOFTWARE',
  SUPPLIES_CONSUMPTION = 'SUPPLIES_CONSUMPTION',
  SUPPLIES_PROTOTYPE = 'SUPPLIES_PROTOTYPE',
  BOOKS_JOURNALS = 'BOOKS_JOURNALS',
  CIVIL_ENGINEERING = 'CIVIL_ENGINEERING',
  // CORRELATED_INFRASTRUCTURE = 'CORRELATED_INFRASTRUCTURE',
  CORRELATED_OTHER = 'CORRELATED_OTHER',
  INSTITUTE_COST = 'INSTITUTE_COST',
  // DONATION = 'DONATION'
}